Cleans up and adds in error handling to API route
