+++ pages/chapters/[name].tsx
