+++ pages/faq/index.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from "next";
-import Header from "components/header";
+import Header from "components/Header";
 import Footer from "components/Footer";
 import FAQ from "components/FAQ";
 const Home: NextPage = () => {
