+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from "next";
 import Head from "next/head";
-import Footer from "components/footer";
-import Header from "components/header";
+import Footer from "components/Footer";
+import Header from "components/Header";
 import MainInfo from "components/MainInfo";
 import Partners from "components/Partners";
