+++ server/actions/Chapter.ts
@@ -27,10 +27,9 @@ export const addChapter = async function (chapterInfo: Chapter) {
     //Define a new chapter model to work with and pass in all the chapterInfo.
     const chapter = new ChapterModel(chapterInfo);
     //Saving the model uploads it to the collection.
-    await chapter.save();
-
-
-    //TODO: Error handling.
+    await chapter.save((err:any) => {
+        if(err) console.log(err);
+    });
 }
 
 export const updateChapter = async function (chapterInfo: Chapter) {
